+++ spec/cose/sign1_spec.rb
@@ -6,6 +6,7 @@ require "cose/algorithm"
 require "cose/error"
 require "cose/key"
 require "cose/sign1"
+require "openssl"
 
 RSpec.describe "COSE::Sign1" do
   context ".deserialize" do
@@ -101,6 +102,122 @@ RSpec.describe "COSE::Sign1" do
           end
         end
       end
+    end
+  end
+
+  context "#sign and #serialize" do
+    {
+      "ES256" => "prime256v1",
+      "ES384" => "secp384r1",
+      "ES512" => "secp521r1"
+    }.each do |alg_name, curve_name|
+      it "signs, serializes, deserializes and verifies a round trip for #{alg_name}" do
+        algorithm = COSE::Algorithm.by_name(alg_name)
+        key = COSE::Key::EC2.from_pkey(OpenSSL::PKey::EC.generate(curve_name))
+
+        sign1 = COSE::Sign1.new(
+          protected_headers: { 1 => algorithm.id },
+          unprotected_headers: {},
+          payload: "This is the content".b
+        )
+
+        expect(sign1.sign(key)).to be(sign1)
+        expect(sign1.signature).to be_a(String)
+
+        deserialized = COSE::Sign1.deserialize(sign1.serialize)
+
+        expect(deserialized.protected_headers).to eq(1 => algorithm.id)
+        expect(deserialized.payload).to eq("This is the content".b)
+        expect(deserialized.verify(key)).to be(true)
+      end
+    end
+
+    if curve_25519_supported?
+      it "signs, serializes, deserializes and verifies a round trip for EdDSA" do
+        algorithm = COSE::Algorithm.by_name("EdDSA")
+        key = COSE::Key::OKP.from_pkey(OpenSSL::PKey.generate_key("ED25519"))
+
+        sign1 = COSE::Sign1.new(
+          protected_headers: { 1 => algorithm.id },
+          unprotected_headers: {},
+          payload: "This is the content".b
+        )
+
+        sign1.sign(key)
+
+        deserialized = COSE::Sign1.deserialize(sign1.serialize)
+
+        expect(deserialized.verify(key)).to be(true)
+      end
+    end
+
+    if rsa_pss_supported?
+      it "signs, serializes, deserializes and verifies a round trip for PS256" do
+        algorithm = COSE::Algorithm.by_name("PS256")
+        key = COSE::Key::RSA.from_pkey(OpenSSL::PKey::RSA.new(2048))
+
+        sign1 = COSE::Sign1.new(
+          protected_headers: { 1 => algorithm.id },
+          unprotected_headers: {},
+          payload: "This is the content".b
+        )
+
+        sign1.sign(key)
+
+        deserialized = COSE::Sign1.deserialize(sign1.serialize)
+
+        expect(deserialized.verify(key)).to be(true)
+      end
+    end
+
+    it "supports signing and verifying with a detached payload" do
+      algorithm = COSE::Algorithm.by_name("ES256")
+      key = COSE::Key::EC2.from_pkey(OpenSSL::PKey::EC.generate("prime256v1"))
+      detached_payload = "Detached content, not embedded in the message".b
+
+      signing_message = COSE::Sign1.new(
+        protected_headers: { 1 => algorithm.id },
+        unprotected_headers: {},
+        payload: detached_payload
+      )
+      signing_message.sign(key)
+
+      sign1 = COSE::Sign1.new(
+        protected_headers: { 1 => algorithm.id },
+        unprotected_headers: {},
+        payload: nil,
+        signature: signing_message.signature
+      )
+
+      deserialized = COSE::Sign1.deserialize(sign1.serialize)
+
+      expect(deserialized.payload).to be_nil
+      expect(deserialized.verify(key, detached_payload: detached_payload)).to be(true)
+      expect { deserialized.verify(key) }.to raise_error(COSE::Error)
+    end
+
+    it "raises COSE::Error when serializing before signing" do
+      sign1 = COSE::Sign1.new(
+        protected_headers: { 1 => COSE::Algorithm.by_name("ES256").id },
+        unprotected_headers: {},
+        payload: "content".b
+      )
+
+      expect { sign1.serialize }.to raise_error(COSE::Error)
+    end
+
+    it "raises COSE::Error when signing with a public-only key" do
+      algorithm = COSE::Algorithm.by_name("ES256")
+      private_key = COSE::Key::EC2.from_pkey(OpenSSL::PKey::EC.generate("prime256v1"))
+      public_key = COSE::Key::EC2.new(crv: private_key.crv, x: private_key.x, y: private_key.y)
+
+      sign1 = COSE::Sign1.new(
+        protected_headers: { 1 => algorithm.id },
+        unprotected_headers: {},
+        payload: "content".b
+      )
+
+      expect { sign1.sign(public_key) }.to raise_error(COSE::Error)
     end
   end
 end